remove duplicate files in rpc,eclair-rpc and replace them with publish testkit dep
remove gen package, fix bugs in testkit generators that weren't properly copied for the gen project
