+++ build.sbt
@@ -44,7 +44,6 @@ lazy val root = project
   .aggregate(
     secp256k1jni,
     core,
-    coreGen,
     coreTest,
     zmq,
     rpc,
@@ -72,21 +71,12 @@ lazy val core = project
     secp256k1jni
   )
 
-lazy val coreGen = project
-  .in(file("core-gen"))
-  .enablePlugins()
-  .settings(commonSettings: _*)
-  .dependsOn(
-    core
-  )
-
 lazy val coreTest = project
   .in(file("core-test"))
   .enablePlugins()
   .settings(commonSettings: _*)
   .dependsOn(
     core,
-    coreGen % "test->test"
   )
 
 lazy val zmq = project
@@ -102,8 +92,7 @@ lazy val rpc = project
   .enablePlugins()
   .settings(commonSettings: _*)
   .dependsOn(
-    core,
-    coreGen % "test->test"
+    core
   )
   .settings(
     testOptions in Test += Tests.Argument("-oF")
