--- core-gen/build.sbt
@@ -1,7 +1,0 @@
-name := "bitcoin-s-core-gen"
-
-libraryDependencies ++= Deps.coreGen
-
-publishArtifact in Compile := false
-
-publishArtifact in Test := true
