+++ testkit/src/main/scala/org/bitcoins/core/gen/ChainParamsGenerator.scala
@@ -1,6 +1,7 @@
 package org.bitcoins.core.gen
 
 import org.bitcoins.core.config._
+import org.bitcoins.core.protocol.ln.LnParams
 import org.scalacheck.Gen
 
 /**
@@ -12,6 +13,12 @@ sealed abstract class ChainParamsGenerator {
 
   def bitcoinNetworkParams: Gen[BitcoinNetwork] =
     Gen.oneOf(MainNet, TestNet3, RegTest)
+
+  def lnNetworkParams: Gen[LnParams] = {
+    Gen.oneOf(LnParams.LnBitcoinMainNet,
+              LnParams.LnBitcoinTestNet,
+              LnParams.LnBitcoinRegTest)
+  }
 }
 
 object ChainParamsGenerator extends ChainParamsGenerator
